add doc strings and type annotations to experiments.py and common.py - classes ExpEnv and ExpOutput had to be moved to prevent circular imports - reformat common.py
