+++ experiments/simbricks/experiment/experiment_output.py
@@ -1,0 +1,62 @@
+# Copyright 2021 Max Planck Institute for Software Systems, and
+# National University of Singapore
+#
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+#
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
+# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+
+import json
+import time
+
+
+class ExpOutput(object):
+    """Manages an experiment's output."""
+
+    def __init__(self, exp):
+        self.exp_name = exp.name
+        self.metadata = exp.metadata
+        self.start_time = None
+        self.end_time = None
+        self.sims = {}
+        self.success = True
+
+    def set_start(self):
+        self.start_time = time.time()
+
+    def set_end(self):
+        self.end_time = time.time()
+
+    def set_failed(self):
+        self.success = False
+
+    def add_sim(self, sim, comp):
+        obj = {
+            'class': sim.__class__.__name__,
+            'cmd': comp.cmd_parts,
+            'stdout': comp.stdout,
+            'stderr': comp.stderr,
+        }
+        self.sims[sim.full_name()] = obj
+
+    def dumps(self):
+        return json.dumps(self.__dict__)
+
+    def loads(self, json_s):
+        for k, v in json.loads(json_s).items():
+            self.__dict__[k] = v
